Prepare pvkgadgets npm package release
